updated run script

+++ env.sh
@@ -5,11 +5,19 @@ DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"
 export CLASSPATH=$(< "$DIR"/classpath)
 
 run() {
-	dataset=$1
+  op=$1
+	dataset=$2
+
+  if [ -z "$dataset" ] || [ -z "$op" ]; then
+    echo "Usage: run operation dataset"
+    return 1
+  fi
+
 	date=$(date +%Y_%m_%d_%H_%M_%S)
 	(
-		mkdir -p outputs/$date
-		cd outputs/$date
+		mkdir -p "$DIR/run/outputs/$date"
+		cd "$DIR/run/outputs/$date"
+    echo "$op $dataset" > info
 
 		(cd ..; rm -f latest; ln -s "$date" latest)
 		for nodes in $(seq "$MAX_NODES" -1 1); do
@@ -18,8 +26,8 @@ run() {
 			mkdir -p "$output_dir"
 
 			P="-agentpath:$HOME/profiler/libasyncProfiler.so=start,alluser,file=$output_dir/profile.%p.svg"
-			cmd="mpirun --oversubscribe -np $nodes java $P com.mycompany.imagej.Main edge_convolution ../../datasets/$dataset ${output}.tiff"
-			(echo $cmd; $cmd < /dev/null) |& tee out
+			cmd="mpirun --oversubscribe -np $nodes java $P com.mycompany.imagej.Main $op  ../../datasets/$dataset $nodes/result.tiff"
+			(echo $cmd; $cmd < /dev/null) |& tee -a out
 		done
 	)
 }
